+++ web/src/App.tsx
@@ -32,6 +32,7 @@ import { InviteAcceptPage } from './pages/InviteAcceptPage';
 import { WorkspacePage } from './pages/WorkspacePage';
 import { SystemPage } from './pages/SystemPage';
 import { RetirementPage } from './pages/RetirementPage';
+import { DebtPayoffPage } from './pages/DebtPayoffPage';
 import { ConnectionsPage } from './pages/ConnectionsPage';
 import { AssistantPage } from './pages/AssistantPage';
 import { SharingPage } from './pages/SharingPage';
@@ -340,6 +341,7 @@ function AuthenticatedApp({
             <NavLink to="/bills">Bills</NavLink>
             <NavLink to="/subscriptions">Subscriptions</NavLink>
             <NavLink to="/retirement">Retirement</NavLink>
+            <NavLink to="/debt-payoff">Debt payoff</NavLink>
           </div>
           <div className="nav-group">
             <div className="nav-group-label">Insights</div>
@@ -395,6 +397,7 @@ function AuthenticatedApp({
           <Route path="/budgets" element={<BudgetsPage />} />
           <Route path="/goals" element={<GoalsPage />} />
           <Route path="/retirement" element={<RetirementPage />} />
+          <Route path="/debt-payoff" element={<DebtPayoffPage />} />
           <Route path="/bills" element={<BillsPage />} />
           <Route path="/subscriptions" element={<SubscriptionsPage />} />
           <Route path="/vehicles" element={<VehiclesPage />} />